+++ components/vf-button/vf-button.react.js
@@ -4,9 +4,9 @@
 // ---
 import React from "react";
 import Fragment from "react-dom-fragment";
-import vfNunjucks from "@visual-framework/vf-extensions-react/assets/nunjucks-slim.js";
-import VfButtonTemplate from "raw-loader!./vf-button.precompiled.js"; // https://webpack.js.org/loaders/raw-loader/
-eval(VfButtonTemplate); // we use eval as we specifically want to run a known template
+// eslint-disable-next-line no-unused-vars
+import VfButtonTemplate from "./vf-button.precompiled.js";  // import templates before the nunjucks env
+import { vfNunjucksEnv } from "@visual-framework/vf-extensions-react/vf-extensions-react.js";
 
 // any JS actions needed on component insertion
 class VfButtonCallback extends React.Component {
@@ -27,7 +27,7 @@ const VfButton = React.memo(({
     React.createElement(Fragment, {
       dangerouslySetInnerHTML: {
         // our HTML is handled by nunjucks, this should not receive user input
-        __html: vfNunjucks.render("vf-button", {
+        __html: vfNunjucksEnv.render("vf-button", {
           text: text, button_href: button_href, theme: theme
         })
       }