+++ tools/vf-component-generator/templates/_component.react.js
@@ -1,12 +1,12 @@
 // <%= componentName %> for React
 // See vf-extensions-react for usage guidance
-// We use vanilla JS templates for react for compaitbility with create react app
+// We use vanilla JS templates for react for compatibility with create react app
 // ---
 import React from "react";
 import Fragment from "react-dom-fragment";
-import vfNunjucks from "@visual-framework/vf-extensions-react/assets/nunjucks-slim.js";
-import <%= componentNameReact %>Template from "raw-loader!./<%= componentName %>.precompiled.js"; // https://webpack.js.org/loaders/raw-loader/
-eval(<%= componentNameReact %>Template); // we use eval as we specifically want to run a known template
+// eslint-disable-next-line no-unused-vars
+import <%= componentNameReact %>Template from "./<%= componentName %>.precompiled.js"; // import templates before the nunjucks env
+import { vfNunjucksEnv } from "@visual-framework/vf-extensions-react/vf-extensions-react.js";
 
 // any JS actions needed on component insertion
 class <%= componentNameReact %>Callback extends React.Component {
@@ -27,7 +27,7 @@ const <%= componentNameReact %> = React.memo(({
       React.createElement(Fragment, {
           dangerouslySetInnerHTML: {
           // our HTML is handled by nunjucks, this should not receive user input
-          __html: vfNunjucks.render('<%= componentName %>', {
+          __html: vfNunjucksEnv.render('<%= componentName %>', {
             variant: variant, newTheme: newTheme, card_image: card_image, card_text: card_text, card_image__alt: card_image__alt, card_title: card_title 
           })
         }