+++ tools/vf-core/gulp-tasks/vf-templates-precompile.js
@@ -1,4 +1,4 @@
-'use strict';
+"use strict";
 
 /**
  * Precompile vf-core Nunjucks templates