+++ tools/vf-extensions-react/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Use react-dom-fragment to return HTML fragments.
   * https://github.com/visual-framework/vf-core/pull/1291
+* Removes no longer needed raw-loader
+* Add vf-extensions-react.js for common Nunjucks environment
+* VF Nunjucks extensions support
 
 ### 0.0.1